+++ configs/recognition/tsm/tsm_r50_mixup_1x1x8_50e_sthv1_rgb.py
@@ -3,7 +3,6 @@ _base_ = [
 ]
 
 # model settings
-# model settings# model settings
 model = dict(
     type='Recognizer2D',
     backbone=dict(